jarvisassitant/requirements.txt
speechrecognition==3.10.0
pyttsx3==2.90
google-generativeai==0.8.5
pyaudio==0.2.14
pygame==2.5.2
pillow==10.1.0
python-dotenv==1.0.0
requests==2.32.0
beautifulsoup4==4.12.0
wikipedia==1.4.0
wolframalpha==5.0.0

